# docs/swagger.yaml
openapi: 3.0.2
info:
  title: Widgets-API
  version: 1.0.0
  description: |-
    ### This is a Swagger UI for our Widgets API


    ### Error handling
    Errors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.

    ```
    {
      "error": "Cleartext explanation of the error that occurred"
    }
    ```

    ### Generic API errors
    These error responses can be returned by any of the described endpoints.

    | HTTP status  |  Description |
    |---|---|
    | 400| Bad input parameters or invalid request. More information can be found in the response body.
    | 401| The consumer does not have access to the requested resource or the provided token is not valid.
    | 404| The requested resource can not be found or the user does not have access to the resource.
    | 5XX| A server error has occurred while processing the request.
paths:
  /widgets:
    get:
      tags:
        - widgets
      summary: Get a list of all widgets
      description: Returns a list of all widgets in API
      responses:
        '200':
          description: Returns a list of widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                List of widgets:
                  value:
                    - Cizzbor
                    - Woowo
                    - Crazlinger
    post:
      tags:
        - new-widget
      summary: Creates a widget
      description: Creates a new widget
      requestBody:
        description: A new widget to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new-widget'
      responses:
        '201':
          description: Widget was created. The Location header points to the new widget.
          headers:
            Location:
              description: URL of the created widget
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/widget'
              examples:
                Created widget:
                  value:
                    id: 6
                    name: Tetris
                    price: 0.99
        '400':
          description: One or all params are missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing params:
                  value:
                    error: One or all params are missing
                Invalid price:
                  value:
                    error: Price must be a number that is 0 or greater
  /widgets/{id}:
    parameters:
      - name: id
        in: path
        description: Id of the widget
        required: true
        schema:
          type: integer
    get:
      tags:
        - details
      summary: Get full details of a widget
      description: Returns full details of a given id widget
      responses:
        '200':
          description: Returns full details of a given id widget
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/widget'
              examples:
                Full details of a given widget:
                  value:
                    id: 1
                    name: Cizzbor
                    price: 29.99
        '400':
          description: Widget id is not a positive integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid id:
                  value:
                    error: Widget id must be a positive integer
        '404':
          description: Widget was not found in widgets array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Widget was not found:
                  value:
                    error: Widget not found
    delete:
      tags:
        - delete-widget
      summary: Deletes a widget
      description: Deletes a widget with the given id
      responses:
        '204':
          description: Widget was deleted
        '400':
          description: Widget id is not a positive integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid id:
                  value:
                    error: Widget id must be a positive integer
        '404':
          description: Widget was not found in widgets array
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Widget was not found:
                  value:
                    error: Widget not found
components:
  schemas:
    error:
      title: Root Type for error
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
    widget:
      title: Root Type for widget
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          format: float
          type: number
      example:
        id: 1
        name: Cizzbor
        price: 29.99
    new-widget:
      title: Root Type for new-widget
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          format: float
          type: number
      example:
        name: Tetris
        price: 0.99
    delete-widget:
      title: Root Type for delete-widget
      type: object
      required:
        - id
      properties:
        id:
          type: integer
      example:
        id: 1